+++ CellFrameDashboardGUI/DiagnosticWorker/AbstractDiagnostic.h
@@ -16,6 +16,7 @@
 #include <QRegularExpression>
 #include <QtMath>
 #include <QSettings>
+#include <QTime>
 
 #include <QDir>
 #include <QNetworkInterface>
@@ -33,7 +34,7 @@ public:
     void set_timeout(int timeout);
     quint64 get_file_size(QString flag, QString path);
     QString get_uptime_string(long sec);
-    QString get_memory_string(long num);
+    QString get_memory_string(size_t num);
     QJsonValue get_mac();
 
     QJsonDocument get_full_info(){return s_full_info;};
